Move GetNewEtcdConfig to utils.go
